PGT&UGT Bot/requirements.txt
aiohttp==3.12.15
discord.py==2.6.4
Jinja2==3.1.6
python-dotenv==1.2.1
